Refactor backup management and remove archiver module
- Updated the PointInTimeMeta structure to use backupmgr package in api_database.go.
- Cleaned up whitespace in SetActionForClusterSettings and GetBackups methods in repmanv3.go.
- Deleted the archiver module as it is no longer needed.
- Introduced a new backupmgr package with comprehensive backup management functionalities, including BackupMetadata, PointInTimeMeta, and various backup strategies.
- Implemented ResticManager for handling backup tasks, including fetching, purging, and managing repository keys.
- Enhanced error handling and logging throughout the backup process.
