--- config/backup.go
@@ -1,82 +1,0 @@
-package config
-
-import (
-	"os"
-	"path/filepath"
-	"time"
-)
-
-type BackupMethod int
-
-const (
-	BackupMethodLogical  = 1
-	BackupMethodPhysical = 2
-)
-
-type BackupStrategy int
-
-const (
-	BackupStrategyFull         = 1
-	BackupStrategyIncremental  = 2
-	BackupStrategyDifferential = 3
-)
-
-type BackupMetadata struct {
-	Id                int64          `json:"id"`
-	StartTime         time.Time      `json:"startTime"`
-	EndTime           time.Time      `json:"endTime"`
-	BackupMethod      BackupMethod   `json:"backupMethod"`
-	BackupTool        string         `json:"backupTool"`
-	BackupToolVersion string         `json:"backupToolVersion"`
-	BackupStrategy    BackupStrategy `json:"backupStrategy"`
-	Source            string         `json:"source"`
-	Dest              string         `json:"dest"`
-	Size              int64          `json:"size"`
-	FileCount         int64          `json:"fileCount"`
-	Compressed        bool           `json:"compressed"`
-	Encrypted         bool           `json:"encrypted"`
-	EncryptionAlgo    string         `json:"encryptionAlgo"`
-	EncryptionKey     string         `json:"encryptionKey"`
-	Checksum          string         `json:"checksum"`
-	RetentionDays     int            `json:"retentionDays"`
-	BinLogFileName    string         `json:"binLogFileName"`
-	BinLogFilePos     uint64         `json:"binLogFilePos"`
-	BinLogGtid        string         `json:"binLogUuid"`
-	Completed         bool           `json:"completed"`
-	SplitUser         bool           `json:"splitUser"`
-	Previous          int64          `json:"previous"`
-}
-
-type PointInTimeMeta struct {
-	IsInPITR    bool
-	UseBinlog   bool
-	Backup      int64
-	RestoreTime int64
-}
-
-func (bm *BackupMetadata) GetSizeAndFileCount() error {
-	var size int64 = 0
-	var fileCount int64 = 0
-	err := filepath.Walk(bm.Dest, func(_ string, info os.FileInfo, err error) error {
-		if err == nil && !info.IsDir() {
-			size += info.Size()
-			fileCount++
-		}
-		return err
-	})
-	bm.Size = size
-	bm.FileCount = fileCount
-	return err
-}
-
-type ReadBinaryLogsBoundary struct {
-	UseTimestamp bool
-	Filename     string
-	Position     int64
-	Timestamp    time.Time
-}
-
-type ReadBinaryLogsRange struct {
-	Start ReadBinaryLogsBoundary
-	End   ReadBinaryLogsBoundary
-}
