+++ config/maps.go
@@ -789,116 +789,6 @@ func FromTasksMap(m *TasksMap, c *TasksMap) *TasksMap {
 	return m
 }
 
-type BackupMetaMap struct {
-	*sync.Map
-}
-
-func NewBackupMetaMap() *BackupMetaMap {
-	s := new(sync.Map)
-	m := &BackupMetaMap{Map: s}
-	return m
-}
-
-func (m *BackupMetaMap) Get(key int64) *BackupMetadata {
-	if v, ok := m.Load(key); ok {
-		return v.(*BackupMetadata)
-	}
-	return nil
-}
-
-func (m *BackupMetaMap) CheckAndGet(key int64) (*BackupMetadata, bool) {
-	v, ok := m.Load(key)
-	if ok {
-		return v.(*BackupMetadata), true
-	}
-	return nil, false
-}
-
-func (m *BackupMetaMap) Set(key int64, value *BackupMetadata) {
-	m.Store(key, value)
-}
-
-func (m *BackupMetaMap) ToNormalMap(c map[int64]*BackupMetadata) {
-	// Clear the old values in the output map
-	for k := range c {
-		delete(c, k)
-	}
-
-	// Insert all values from the BackupMetaMap to the output map
-	m.Callback(func(key int64, value *BackupMetadata) bool {
-		c[key] = value
-		return true
-	})
-}
-
-func (m *BackupMetaMap) ToNewMap() map[int64]*BackupMetadata {
-	result := make(map[int64]*BackupMetadata)
-	m.Range(func(k, v any) bool {
-		result[k.(int64)] = v.(*BackupMetadata)
-		return true
-	})
-	return result
-}
-
-func (m *BackupMetaMap) Callback(f func(key int64, value *BackupMetadata) bool) {
-	m.Range(func(k, v any) bool {
-		return f(k.(int64), v.(*BackupMetadata))
-	})
-}
-
-func (m *BackupMetaMap) Clear() {
-	m.Range(func(key, value any) bool {
-		m.Delete(key.(int64))
-		return true
-	})
-}
-
-func FromNormalBackupMetaMap(m *BackupMetaMap, c map[int64]*BackupMetadata) *BackupMetaMap {
-	if m == nil {
-		m = NewBackupMetaMap()
-	} else {
-		m.Clear()
-	}
-
-	for k, v := range c {
-		m.Set(k, v)
-	}
-
-	return m
-}
-
-func FromBackupMetaMap(m *BackupMetaMap, c *BackupMetaMap) *BackupMetaMap {
-	if m == nil {
-		m = NewBackupMetaMap()
-	} else {
-		m.Clear()
-	}
-
-	if c != nil {
-		c.Callback(func(key int64, value *BackupMetadata) bool {
-			m.Set(key, value)
-			return true
-		})
-	}
-
-	return m
-}
-
-// GetBackupsByToolAndSource retrieves backups with the same backupTool and source.
-func (b *BackupMetaMap) GetPreviousBackup(backupTool string, source string) *BackupMetadata {
-	var result *BackupMetadata
-	b.Map.Range(func(key, value interface{}) bool {
-		if backup, ok := value.(*BackupMetadata); ok {
-			if backup.BackupTool == backupTool && backup.Source == source {
-				result = backup
-				return false
-			}
-		}
-		return true
-	})
-	return result
-}
-
 type VersionsMap struct {
 	*sync.Map
 }
